Add mutual room binding.
Now when a exit is configured for one room, the other room automatically binds its
corresponding exit to this room.
Also reverted renaming of `setExit()`, which happened in
cc810c392d7d510347e36cd0989ad0607180d47c
So now it is back to `setExit()`

Also added itemId for Appliance.

Updated Inventory to use `this` keyword.